+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 * Unreleased
+    * Add `examples/*.py` to demonstrate how to use `SchemaGenerator` as a
+      library.
 * 1.5 (2021-11-14)
     * Make the column order in the BQ schema file match the order of appearance
       in the JSON data file using the `--preserve_input_sort_order` flag.